// apps/amplify-be/src/graphql/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type User = {
  __typename: "User",
  username: string,
  email: string,
  profile?: Profile | null,
  notifications?: ModelNotificationConnection | null,
  createdAt: string,
  updatedAt: string,
  userProfileUsername?: string | null,
};

export type Profile = {
  __typename: "Profile",
  username: string,
  bio?: string | null,
  user?: User | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelNotificationConnection = {
  __typename: "ModelNotificationConnection",
  items:  Array<Notification | null >,
  nextToken?: string | null,
};

export type Notification = {
  __typename: "Notification",
  id: string,
  username: string,
  title: string,
  message: string,
  read: boolean,
  user?: User | null,
  createdAt: string,
  updatedAt: string,
  userNotificationsUsername?: string | null,
};
// ...
export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export enum TransferStatus {
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
  REJECTED = "REJECTED",
  PENDING = "PENDING",
}


export type ModelTransferConditionInput = {
  from?: ModelStringInput | null,
  to?: ModelStringInput | null,
  amount?: ModelFloatInput | null,
  status?: ModelTransferStatusInput | null,
  and?: Array< ModelTransferConditionInput | null > | null,
  or?: Array< ModelTransferConditionInput | null > | null,
  not?: ModelTransferConditionInput | null,
};

export type ModelFloatInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelTransferStatusInput = {
  eq?: TransferStatus | null,
  ne?: TransferStatus | null,
};

export type Transfer = {
  __typename: "Transfer",
  id: string,
  from: string,
  to: string,
  amount: number,
  status: TransferStatus,
  createdAt: string,
  updatedAt: string,
};

export type UpdateTransferInput = {
  id: string,
  from?: string | null,
  to?: string | null,
  amount?: number | null,
  status?: TransferStatus | null,
};

export type DeleteTransferInput = {
  id: string,
};

export type ModelUserFilterInput = {
  username?: ModelStringInput | null,
  email?: ModelStringInput | null,
  and?: Array< ModelUserFilterInput | null > | null,
  or?: Array< ModelUserFilterInput | null > | null,
  not?: ModelUserFilterInput | null,
  userProfileUsername?: ModelStringInput | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}


export type ModelUserConnection = {
  __typename: "ModelUserConnection",
  items:  Array<User | null >,
  nextToken?: string | null,
};

export type ModelProfileFilterInput = {
  username?: ModelStringInput | null,
  bio?: ModelStringInput | null,
  and?: Array< ModelProfileFilterInput | null > | null,
  or?: Array< ModelProfileFilterInput | null > | null,
  not?: ModelProfileFilterInput | null,
};

export type ModelProfileConnection = {
  __typename: "ModelProfileConnection",
  items:  Array<Profile | null >,
  nextToken?: string | null,
};

export type ModelNotificationFilterInput = {
  id?: ModelIDInput | null,
  username?: ModelStringInput | null,
  title?: ModelStringInput | null,
  message?: ModelStringInput | null,
  read?: ModelBooleanInput | null,
  and?: Array< ModelNotificationFilterInput | null > | null,
  or?: Array< ModelNotificationFilterInput | null > | null,
  not?: ModelNotificationFilterInput | null,
  userNotificationsUsername?: ModelStringInput | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelTransferFilterInput = {
  id?: ModelIDInput | null,
  from?: ModelStringInput | null,
  to?: ModelStringInput | null,
  amount?: ModelFloatInput | null,
  status?: ModelTransferStatusInput | null,
  and?: Array< ModelTransferFilterInput | null > | null,
  or?: Array< ModelTransferFilterInput | null > | null,
  not?: ModelTransferFilterInput | null,
};

export type ModelTransferConnection = {
  __typename: "ModelTransferConnection",
  items:  Array<Transfer | null >,
  nextToken?: string | null,
};
// ...
export type GetUserQueryVariables = {
  username: string,
};

export type GetUserQuery = {
  getUser?:  {
    __typename: "User",
    username: string,
    email: string,
    profile?:  {
      __typename: "Profile",
      username: string,
      bio?: string | null,
      user?:  {
        __typename: "User",
        username: string,
        email: string,
        createdAt: string,
        updatedAt: string,
        userProfileUsername?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    notifications?:  {
      __typename: "ModelNotificationConnection",
      items:  Array< {
        __typename: "Notification",
        id: string,
        username: string,
        title: string,
        message: string,
        read: boolean,
        createdAt: string,
        updatedAt: string,
        userNotificationsUsername?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userProfileUsername?: string | null,
  } | null,
};

export type ListUsersQueryVariables = {
  username?: string | null,
  filter?: ModelUserFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListUsersQuery = {
  listUsers?:  {
    __typename: "ModelUserConnection",
    items:  Array< {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetProfileQueryVariables = {
  username: string,
};

export type GetProfileQuery = {
  getProfile?:  {
    __typename: "Profile",
    username: string,
    bio?: string | null,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type ListProfilesQueryVariables = {
  username?: string | null,
  filter?: ModelProfileFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListProfilesQuery = {
  listProfiles?:  {
    __typename: "ModelProfileConnection",
    items:  Array< {
      __typename: "Profile",
      username: string,
      bio?: string | null,
      user?:  {
        __typename: "User",
        username: string,
        email: string,
        createdAt: string,
        updatedAt: string,
        userProfileUsername?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetNotificationQueryVariables = {
  id: string,
};

export type GetNotificationQuery = {
  getNotification?:  {
    __typename: "Notification",
    id: string,
    username: string,
    title: string,
    message: string,
    read: boolean,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userNotificationsUsername?: string | null,
  } | null,
};

export type ListNotificationsQueryVariables = {
  filter?: ModelNotificationFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListNotificationsQuery = {
  listNotifications?:  {
    __typename: "ModelNotificationConnection",
    items:  Array< {
      __typename: "Notification",
      id: string,
      username: string,
      title: string,
      message: string,
      read: boolean,
      user?:  {
        __typename: "User",
        username: string,
        email: string,
        createdAt: string,
        updatedAt: string,
        userProfileUsername?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userNotificationsUsername?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetTransferQueryVariables = {
  id: string,
};

export type GetTransferQuery = {
  getTransfer?:  {
    __typename: "Transfer",
    id: string,
    from: string,
    to: string,
    amount: number,
    status: TransferStatus,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type ListTransfersQueryVariables = {
  filter?: ModelTransferFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListTransfersQuery = {
  listTransfers?:  {
    __typename: "ModelTransferConnection",
    items:  Array< {
      __typename: "Transfer",
      id: string,
      from: string,
      to: string,
      amount: number,
      status: TransferStatus,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};